feat(serve): encoder_attention, and tests for the three encoder ops
Non-causal GQA over a whole sequence: two strided-batched cuBLASLt products
around a masked softmax. With three query heads over at most a couple of
thousand tokens the score matrix is tens of megabytes, so materialising it costs
less than a hand-written flash kernel would cost to get right, and the second
product then reads it as an ordinary matrix. Q, K and V are read straight out of
the fused projection by pointer offset and leading dimension; nothing is split
out first. Scores are held column-major as s[key, query] so the softmax reduces
over stride-1 data and the same buffer feeds the second product untransposed.

The window is symmetric — abs(i - j) < W — because that is what a window means
once attention is bidirectional. The two reference implementations disagree here
(transformers uses W, llama.cpp's SWA_TYPE_SYMMETRIC uses W/2); this follows
transformers, and the header says why.

Five cases, including EmbeddingGemma's own geometry at 600 tokens against a
512 window, where some rows clip at both ends and some at neither, and a window
wider than the sequence, which must agree with the global case since a mask that
admits everything is no mask.

mean_pool changed contract while being tested: its output is now FP32. The
chunked case failed at 0.36%, and the cause was the design rather than the
tolerance — accumulating a running sum through BF16 storage rounds it once per
chunk, which is an order of magnitude worse than the quantised weights it is
summing. With FP32 output the criterion tightens 10x, to 2e-4.

+++ csrc/src/serve/ops/kernel/mean_pool.cuh
@@ -15,7 +15,7 @@ inline constexpr int kMeanPoolBlock = 256;
 // `scale` is 1/count for a mean and 1.0 for a running sum; `accumulate` adds
 // into out rather than overwriting, so a chunked prefill can pool across calls.
 __launch_bounds__(kMeanPoolBlock) __global__
-    void mean_pool_kernel(const __nv_bfloat16* __restrict__ x, __nv_bfloat16* __restrict__ out,
+    void mean_pool_kernel(const __nv_bfloat16* __restrict__ x, float* __restrict__ out,
                           std::int32_t hidden, std::int32_t count, float scale, bool accumulate) {
     const std::int32_t row = blockIdx.x;
     if (row >= hidden) { return; }
@@ -35,7 +35,9 @@ __launch_bounds__(kMeanPoolBlock) __global__
 
     if (threadIdx.x == 0) {
         const float value = partial[0] * scale;
-        out[row] = __float2bfloat16(accumulate ? __bfloat162float(out[row]) + value : value);
+        // FP32 storage: a running sum rounded to BF16 once per chunk drifts far
+        // more than the quantised weights it is summing.
+        out[row] = accumulate ? out[row] + value : value;
     }
 }
