+++ csrc/src/serve/ops/launcher/mean_pool.cu
@@ -17,7 +17,7 @@ void mean_pool_launch(const Tensor& x, int count, bool accumulate, Tensor& out,
     // the last chunk has landed.
     const float scale = accumulate ? 1.0f : 1.0f / static_cast<float>(count);
     mean_pool_kernel<<<static_cast<unsigned int>(hidden), kMeanPoolBlock, 0, stream>>>(
-        static_cast<const __nv_bfloat16*>(x.data), static_cast<__nv_bfloat16*>(out.data), hidden,
+        static_cast<const __nv_bfloat16*>(x.data), static_cast<float*>(out.data), hidden,
         static_cast<std::int32_t>(count), scale, accumulate);
     CUDA_CHECK(cudaGetLastError());
 }
